requirements.txt
requests~=2.25.1
pandas~=2.1.4
infi.clickhouse-orm